polish finish 1

+++ libftprintf.h
@@ -6,14 +6,16 @@
 /*   By: tjerdnap <tjerdnap@student.42bangkok.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/29 14:12:06 by tjerdnap          #+#    #+#             */
-/*   Updated: 2024/03/04 14:25:28 by tjerdnap         ###   ########.fr       */
+/*   Updated: 2024/03/04 15:51:49 by tjerdnap         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFTPRINTF_H
 # define LIBFTPRINTF_H
 
+# include <limits.h>
 # include <stdarg.h>
+# include <stdint.h>
 # include <stdio.h>
 # include <stdlib.h>
 # include <unistd.h>
@@ -21,7 +23,7 @@
 int	ft_print_char(char c);
 int	ft_print_str(char *s);
 int	ft_print_int(int num);
-int	ft_print_hex(size_t hex, int lower);
+int	ft_print_hex(unsigned long long hex, int lower);
 int	ft_print_ptr(void *ptr);
 int	ft_print_uint(unsigned int num);
 int	ft_find_format(const char *format, va_list ap, int j);
